src/multistream_yolo/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
flask-restx==1.2.0
matplotlib==3.5.1
netifaces==0.11.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
Pillow>=9.5.0
ultralytics==8.0.203
